requirements.txt
# Core dependencies
openai>=1.12.0
python-dotenv>=1.0.0
chromadb>=0.4.22

# Wikipedia API
wikipedia-api>=0.6.0

# Optional but recommended
numpy>=1.24.0
requests>=2.31.0
